+++ internal/pkg/info_user/repository.go
@@ -1,4 +1,4 @@
-package info
+package info_user
 
 import dataStruct "github.com/go-park-mail-ru/2023_1_MRGA.git/internal/app/data_struct"
